Split Serialization to smaller chunks.
