--- trace/src/main/org/jetbrains/lincheck/trace/Serialization.kt
@@ -1,1145 +1,0 @@
-/*
- * Lincheck
- *
- * Copyright (C) 2019 - 2025 JetBrains s.r.o.
- *
- * This Source Code Form is subject to the terms of the
- * Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
- * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-package org.jetbrains.lincheck.trace
-
-import org.jetbrains.lincheck.descriptors.AccessLocation
-import org.jetbrains.lincheck.descriptors.AccessPath
-import org.jetbrains.lincheck.descriptors.ArrayElementByNameAccessLocation
-import org.jetbrains.lincheck.descriptors.ClassDescriptor
-import org.jetbrains.lincheck.descriptors.FieldDescriptor
-import org.jetbrains.lincheck.descriptors.MethodDescriptor
-import org.jetbrains.lincheck.descriptors.VariableDescriptor
-import org.jetbrains.lincheck.util.Logger
-import java.io.*
-import java.nio.ByteBuffer
-import java.nio.file.Files
-import java.util.concurrent.ArrayBlockingQueue
-import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.atomic.AtomicInteger
-import java.util.concurrent.atomic.AtomicLong
-
-import java.util.zip.ZipEntry
-import java.util.zip.ZipOutputStream
-import kotlin.io.path.Path
-import kotlin.math.absoluteValue
-import kotlin.math.max
-
-// Buffer for saving trace in one piece
-private const val OUTPUT_BUFFER_SIZE: Int = 16 * 1024 * 1024
-
-// 1 MiB
-private const val PER_THREAD_DATA_BUFFER_SIZE: Int = 1024 * 1024
-
-// When flush per-thread data block?
-private const val DATA_FLUSH_THRESHOLD: Int = 1024
-
-/**
- * It is a strategy to collect trace: it can be full-track in memory or streaming to a file on-the-fly
- */
-interface TraceCollectingStrategy {
-    /**
-     * Register current thread in strategy.
-     */
-    fun registerCurrentThread(threadId: Int)
-
-    /**
-     * Makes sure that thread has written all of its recorded data.
-     */
-    fun completeThread(thread: Thread)
-
-    /**
-     * Must be called when a new tracepoint is created.
-     *
-     * @param parent Current top of the call stack, if exists.
-     * @param created New tracepoint
-     */
-    fun tracePointCreated(parent: TRContainerTracePoint?, created: TRTracePoint)
-
-    /**
-     * Must be called when the container trace point is ended and popped from the trace tree.
-     *
-     * @param thread thread of the [container] trace point.
-     * @param container the completed container trace point.
-     */
-    fun completeContainerTracePoint(thread: Thread, container: TRContainerTracePoint)
-
-    /**
-     * Must be called when trace is finished
-     */
-    fun traceEnded()
-}
-
-
-/**
- * It is a strategy to save one tracepoint.
- */
-internal interface TraceWriter : DataOutput, Closeable {
-    /**
-     * Saves dependencies of [TRObject], if needed.
-     * This must be called before [startWriteAnyTracepoint] for all used [TRObject]s.
-     */
-    fun preWriteTRObject(value: TRObject?)
-
-    /**
-     * Saves [TRObject] itself.
-     * Must be called after [startWriteAnyTracepoint] or [startWriteContainerTracepointFooter].
-     */
-    fun writeTRObject(value: TRObject?)
-
-    /**
-     * Marks the beginning of a tracepoint (before the first byte of tracepoint is written).
-     */
-    fun startWriteAnyTracepoint()
-
-    /**
-     * Marks the end of the leaf (fix-sized) tracepoint.
-     */
-    fun endWriteLeafTracepoint()
-
-    /**
-     * Mark the end of the container tracepoint's header (now only TRMethodCallTracepoint is a container one).
-     */
-    fun endWriteContainerTracepointHeader(id: Int)
-
-    /**
-     * Mark the beginning of container tracepoint's footer (After all children are saved).
-     */
-    fun startWriteContainerTracepointFooter()
-
-    /**
-     * Marks the end of container tracepoint's footer.
-     */
-    fun endWriteContainerTracepointFooter(id: Int)
-
-    /**
-     * Write [name] of the thread.
-     * This must be called before [startWriteAnyTracepoint] or [startWriteContainerTracepointFooter] for all thread names.
-     */
-    fun writeThreadName(id: Int, name: String)
-
-    /**
-     * Write [ClassDescriptor] from context referred by given `id`, if needed.
-     * This must be called before [startWriteAnyTracepoint] or [startWriteContainerTracepointFooter] for all used class descriptors.
-     */
-    fun writeClassDescriptor(id: Int)
-
-    /**
-     * Write [MethodDescriptor] from context referred by given `id`, if needed.
-     * This must be called before [startWriteAnyTracepoint] or [startWriteContainerTracepointFooter] for all used method descriptors.
-     */
-    fun writeMethodDescriptor(id: Int)
-
-    /**
-     * Write [FieldDescriptor] from context referred by given `id`, if needed.
-     * This must be called before [startWriteAnyTracepoint] or [startWriteContainerTracepointFooter] for all used field descriptors.
-     */
-    fun writeFieldDescriptor(id: Int)
-
-    /**
-     * Write [VariableDescriptor] from context referred by given `id` if needed.
-     * This must be called before [startWriteAnyTracepoint] or [startWriteContainerTracepointFooter] for all used variable descriptors.
-     */
-    fun writeVariableDescriptor(id: Int)
-
-    /**
-     * Write [StackTraceElement] from context referred by given code location `id`, if needed.
-     * This must be called before [startWriteAnyTracepoint] or [startWriteContainerTracepointFooter] for all used code locations.
-     */
-    fun writeCodeLocation(id: Int)
-}
-
-/**
- * Interface to check and mark if a given piece of reference data was already stored.
- */
-internal interface ContextSavingState {
-    fun isClassDescriptorSaved(id: Int): Boolean
-    fun markClassDescriptorSaved(id: Int)
-    fun isMethodDescriptorSaved(id: Int): Boolean
-    fun markMethodDescriptorSaved(id: Int)
-    fun isFieldDescriptorSaved(id: Int): Boolean
-    fun markFieldDescriptorSaved(id: Int)
-    fun isVariableDescriptorSaved(id: Int): Boolean
-    fun markVariableDescriptorSaved(id: Int)
-    fun isCodeLocationSaved(id: Int): Boolean
-    fun markCodeLocationSaved(id: Int)
-
-    /**
-     * Return positive string id, if it was stored already, and negative, if it should be stored
-     * with absolute value of this id.
-     */
-    fun isStringSaved(value: String): Int
-
-    /**
-     * Mark string as stored. Do nothing if it was not passed to [isStringSaved].
-     */
-    fun markStringSaved(value: String)
-
-    /**
-     * Return positive access path id, if it was stored already, and negative, if it should be stored
-     * with absolute value of this id.
-     */
-    fun isAccessPathSaved(value: AccessPath): Int
-
-    /**
-     * Mark access path as stored. Do nothing if it was not passed to [isAccessPathSaved].
-     */
-    fun markAccessPathSaved(value: AccessPath)
-}
-
-/**
- * [dataStream] responsible for operations like `close()` and [dataOutput] for real data output.
- *
- * As this class is used with both JDK's [DataOutputStream] and project-local [ByteBufferOutputStream],
- * and [OutputStream] is abstract class and not an interface, it is impossible to make one property which
- * is compatible with both [DataOutputStream] and [ByteBufferOutputStream] at the same time.
- *
- * `dataStream` can be relaxed to [Closeable], but it will hide its intention even more.
- */
-private sealed class TraceWriterBase(
-    private val context: TraceContext,
-    private val contextState: ContextSavingState,
-    protected val dataStream: OutputStream,
-    protected val dataOutput: DataOutput
-): TraceWriter, DataOutput by dataOutput {
-    // Stack of "container" tracepoints
-    private val containerStack = mutableListOf<Pair<Int, Long>>()
-
-    private var inTracepointBody = false
-    private var footerPosition: Long = -1
-
-    protected abstract val currentDataPosition: Long
-    abstract val writerId: Int
-
-    override fun close() {
-        dataOutput.writeKind(ObjectKind.EOF)
-        dataStream.close()
-
-        writeIndexCell(ObjectKind.EOF,-1, -1, -1)
-    }
-
-    override fun preWriteTRObject(value: TRObject?) {
-        check(!inTracepointBody) { "Cannot write TRObject dependency into tracepoint body" }
-        if (value == null || value.isPrimitive || value.isSpecial) return
-        writeClassDescriptor(value.classNameId)
-    }
-
-    override fun writeTRObject(value: TRObject?) {
-        check(inTracepointBody) { "Cannot write TRObject outside tracepoint body" }
-        dataOutput.writeTRObject(value)
-    }
-
-    override fun startWriteAnyTracepoint() {
-        check(!inTracepointBody) { "Cannot start nested tracepoint body" }
-        dataOutput.writeKind(ObjectKind.TRACEPOINT)
-        inTracepointBody = true
-    }
-
-    override fun endWriteLeafTracepoint() {
-        check(inTracepointBody) { "Cannot end tracepoint body not in tracepoint" }
-        inTracepointBody = false
-    }
-
-    override fun endWriteContainerTracepointHeader(id: Int) {
-        check(inTracepointBody) { "Cannot end tracepoint header not in tracepoint" }
-        inTracepointBody = false
-
-        // Store where container content starts
-        containerStack.add(id to currentDataPosition)
-    }
-
-    override fun startWriteContainerTracepointFooter() {
-        check(!inTracepointBody) { "Cannot start nested tracepoint footer" }
-        inTracepointBody = true
-        footerPosition = currentDataPosition
-
-        check(containerStack.isNotEmpty()) {
-            "Calls endWriteContainerTracepointHeader() / startWriteContainerTracepointFooter() are not balanced"
-        }
-
-        // Start object
-        dataOutput.writeKind(ObjectKind.TRACEPOINT_FOOTER)
-    }
-
-    override fun endWriteContainerTracepointFooter(id: Int) {
-        check(inTracepointBody) { "Cannot end tracepoint footer not in tracepoint" }
-        check(footerPosition >= 0) { "Cannot end tracepoint footer not in tracepoint footer" }
-
-        val (storedId, startPos) = containerStack.removeLast()
-        check(id == storedId) {
-            "Calls endWriteContainerTracepointHeader($storedId) / startWriteContainerTracepointFooter($id) are not balanced"
-        }
-        writeIndexCell(ObjectKind.TRACEPOINT, id, startPos, footerPosition)
-
-        inTracepointBody = false
-        footerPosition = -1
-    }
-
-    override fun writeThreadName(id: Int, name: String) {
-        check(!inTracepointBody) { "Cannot write thread name inside tracepoint" }
-
-        val position = currentDataPosition
-        dataOutput.writeKind(ObjectKind.THREAD_NAME)
-        dataOutput.writeInt(id)
-        dataOutput.writeUTF(name)
-        writeIndexCell(ObjectKind.THREAD_NAME, id, position, -1)
-    }
-
-    override fun writeClassDescriptor(id: Int) {
-        check(!inTracepointBody) { "Cannot save reference data inside tracepoint" }
-        if (contextState.isClassDescriptorSaved(id)) return
-        // Write class descriptor into data and position into index
-        val position = currentDataPosition
-        dataOutput.writeKind(ObjectKind.CLASS_DESCRIPTOR)
-        dataOutput.writeInt(id)
-        dataOutput.writeClassDescriptor(context.getClassDescriptor(id))
-        contextState.markClassDescriptorSaved(id)
-
-        writeIndexCell(ObjectKind.CLASS_DESCRIPTOR, id, position, -1)
-    }
-
-    override fun writeMethodDescriptor(id: Int) {
-        check(!inTracepointBody) { "Cannot save reference data inside tracepoint" }
-        if (contextState.isMethodDescriptorSaved(id)) return
-        val descriptor = context.getMethodDescriptor(id)
-        writeClassDescriptor(descriptor.classId)
-
-        // Write method descriptor into data and position into index
-        val position = currentDataPosition
-        dataOutput.writeKind(ObjectKind.METHOD_DESCRIPTOR)
-        dataOutput.writeInt(id)
-        dataOutput.writeMethodDescriptor(descriptor)
-        contextState.markMethodDescriptorSaved(id)
-
-        writeIndexCell(ObjectKind.METHOD_DESCRIPTOR, id, position, -1)
-    }
-
-    override fun writeFieldDescriptor(id: Int) {
-        check(!inTracepointBody) { "Cannot save reference data inside tracepoint" }
-        if (contextState.isFieldDescriptorSaved(id)) return
-        val descriptor = context.getFieldDescriptor(id)
-        writeClassDescriptor(descriptor.classId)
-        // Write field descriptor into data and position into index
-        val position = currentDataPosition
-        dataOutput.writeKind(ObjectKind.FIELD_DESCRIPTOR)
-        dataOutput.writeInt(id)
-        dataOutput.writeFieldDescriptor(descriptor)
-        contextState.markFieldDescriptorSaved(id)
-
-        writeIndexCell(ObjectKind.FIELD_DESCRIPTOR, id, position, -1)
-    }
-
-    override fun writeVariableDescriptor(id: Int) {
-        check(!inTracepointBody) { "Cannot save reference data inside tracepoint" }
-        if (contextState.isVariableDescriptorSaved(id)) return
-        // Write variable descriptor into data and position into index
-        val position = currentDataPosition
-        dataOutput.writeKind(ObjectKind.VARIABLE_DESCRIPTOR)
-        dataOutput.writeInt(id)
-        dataOutput.writeVariableDescriptor(context.getVariableDescriptor(id))
-        contextState.markVariableDescriptorSaved(id)
-
-        writeIndexCell(ObjectKind.VARIABLE_DESCRIPTOR, id, position, -1)
-    }
-
-    override fun writeCodeLocation(id: Int) {
-        check(!inTracepointBody) {
-            "Cannot save reference data inside tracepoint"
-        }
-        if (id == UNKNOWN_CODE_LOCATION_ID) return
-        if (contextState.isCodeLocationSaved(id)) return
-
-        val codeLocation = context.stackTrace(id)
-        val accessPath = context.accessPath(id)
-        val argumentNames = context.methodCallArgumentNames(id)
-        // All strings only once. It will have duplications with class and method descriptors,
-        // but size loss is negligible and this way is simpler
-        val fileNameId = writeString(codeLocation.fileName)
-        val classNameId = writeString(codeLocation.className)
-        val methodNameId = writeString(codeLocation.methodName)
-        val accessPathId = writeAccessPath(accessPath)
-        val argumentNamesIds = argumentNames?.map { writeAccessPath(it) }
-
-        // Code location into data and position into index
-        val position = currentDataPosition
-        dataOutput.writeKind(ObjectKind.CODE_LOCATION)
-        dataOutput.writeInt(id)
-        dataOutput.writeInt(fileNameId)
-        dataOutput.writeInt(classNameId)
-        dataOutput.writeInt(methodNameId)
-        dataOutput.writeInt(codeLocation.lineNumber)
-        dataOutput.writeInt(accessPathId)
-        dataOutput.writeInt(argumentNamesIds?.size ?: 0)
-        argumentNamesIds?.forEach { dataOutput.writeInt(it) }
-        contextState.markCodeLocationSaved(id)
-
-        writeIndexCell(ObjectKind.CODE_LOCATION, id, position, -1)
-    }
-
-    protected open fun writeString(value: String?): Int {
-        check(!inTracepointBody) { "Cannot save reference data inside tracepoint" }
-        if (value == null) return -1
-
-        val id = contextState.isStringSaved(value)
-        if (id > 0) return id
-
-        val position = currentDataPosition
-        dataOutput.writeKind(ObjectKind.STRING)
-        dataOutput.writeInt(-id)
-        dataOutput.writeUTF(value)
-        contextState.markStringSaved(value)
-
-        // It cannot fail
-        writeIndexCell(ObjectKind.STRING, -id, position, -1)
-
-        return -id
-    }
-
-    /**
-     * Writes access path [value] to the output stream.
-     *
-     * The method gets an order of all access paths that are inside the [value], in which
-     * they should be serialized (innermost go first -- top-sort).
-     * After that method first serializes all dependencies of the all access locations inside
-     * each access path and then saves access paths in top-sort order.
-     *
-     * ```
-     * first: [variable descriptor 1] [field descriptor 1] [variable descriptor 2] ...
-     * then: [ACCESS_LOCATION] [id 1] [locations count] [location type] [data] [location type] [data] ...
-     *       [ACCESS_LOCATION] [id 2] [locations count] [location type] [data] ...
-     * ```
-     *
-     * Such order is required, because [AccessPath] is a recursive structure, which may contain another access paths inside.
-     * They should come first in the serialization order for easier deserialization later. So when we need to construct
-     * an [AccessLocation] which expects [AccessPath] as an argument, we would be sure that it is
-     * present in the trace context and can be retrieved via id. So such locations are serialized the folowwing way:
-     * ```
-     * [location type] [another access path id]
-     * ```
-     *
-     * Also, each location inside access path may contain variable/field descriptors, which also should be
-     * serialized beforehand for easier deserialization later. Their structure looks similar way:
-     * ```
-     * [location type] [field/variable descriptor id]
-     * ```
-     */
-    private fun writeAccessPath(value: AccessPath?): Int {
-        check(!inTracepointBody) { "Cannot save reference data inside tracepoint" }
-        if (value == null) return -1
-
-        val savingOrder = collectAccessPathsInSavingOrder(value)
-        val id = writeAccessPaths(value, savingOrder)
-        return id
-    }
-
-    private fun writeAccessPaths(root: AccessPath, savingOrder: List<AccessPath>): Int {
-        var rootId = 0
-
-        savingOrder
-            // first, we save all references of every location inside each access path
-            .onEach { value ->
-                value.locations.forEach { location ->
-                    location.saveReferences(this, context)
-                }
-            }
-            // then, save the access paths in correct order
-            .onEach { value ->
-                val position = currentDataPosition
-                val id = contextState.isAccessPathSaved(value)
-                if (value == root) rootId = id.absoluteValue
-                if (id > 0) return@onEach // already saved
-
-                dataOutput.writeKind(ObjectKind.ACCESS_PATH)
-                dataOutput.writeInt(-id)
-                dataOutput.writeInt(value.locations.size)
-
-                value.locations.forEach { location ->
-                    location.save(this, context, contextState)
-                }
-
-                contextState.markAccessPathSaved(value)
-                writeIndexCell(ObjectKind.ACCESS_PATH, -id, position, -1)
-            }
-
-        check(rootId > 0) { "Root access path $root was not added to the saved access paths: $savingOrder" }
-        return rootId
-    }
-
-    /**
-     * @return all `AccessPath`s, reachable from [value], in top-sort order (from innermost to outermost).
-     */
-    private fun collectAccessPathsInSavingOrder(value: AccessPath): List<AccessPath> {
-        val order = mutableListOf<AccessPath>()
-        collectAccessPathsInSavingOrder(value, mutableSetOf(), order)
-        return order
-    }
-
-    private fun collectAccessPathsInSavingOrder(current: AccessPath, visited: MutableSet<AccessPath>, order: MutableList<AccessPath>) {
-        visited.add(current)
-        current.locations.forEach { location ->
-            if (location is ArrayElementByNameAccessLocation && !visited.contains(location.indexAccessPath)) {
-                collectAccessPathsInSavingOrder(location.indexAccessPath, visited, order)
-            }
-        }
-        order.add(current)
-    }
-
-    protected fun resetTracepointState() {
-        inTracepointBody = false
-        footerPosition = -1
-    }
-
-
-    protected abstract fun writeIndexCell(kind: ObjectKind, id: Int, startPos: Long, endPos: Long)
-}
-
-internal data class IndexCell(
-    val kind: ObjectKind,
-    val id: Int,
-    val startPos: Long,
-    val endPos: Long
-)
-
-private interface BlockSaver {
-    fun saveDataAndIndexBlock(writerId: Int, logicalBlockStart: Long, dataBlock: ByteBuffer, indexList: List<IndexCell>)
-}
-
-
-// Leave some space for metadata
-private const val MAX_STRING_SIZE = PER_THREAD_DATA_BUFFER_SIZE - 1024
-
-private class BufferedTraceWriter (
-    override val writerId: Int,
-    context: TraceContext,
-    contextState: ContextSavingState,
-    private val storage: BlockSaver,
-    private val bufferStream: ByteBufferOutputStream = ByteBufferOutputStream(PER_THREAD_DATA_BUFFER_SIZE)
-) : TraceWriterBase(
-    context = context,
-    contextState = contextState,
-    dataStream = bufferStream,
-    dataOutput = bufferStream
-) {
-    private var currentStartDataPosition: Long = 0
-    private var index = mutableListOf<IndexCell>()
-
-    override val currentDataPosition: Long get() = currentStartDataPosition + bufferStream.position()
-
-    override fun close() {
-        flush()
-        super.close()
-    }
-
-    override fun endWriteLeafTracepoint() {
-        super.endWriteLeafTracepoint()
-        maybeFlushData()
-    }
-
-    override fun endWriteContainerTracepointFooter(id: Int) {
-        super.endWriteContainerTracepointFooter(id)
-        maybeFlushData()
-    }
-
-    override fun writeIndexCell(kind: ObjectKind, id: Int, startPos: Long, endPos: Long) {
-        index.add(IndexCell(kind, id, startPos, endPos))
-        // Rollback to here in case of overflow, as the index cell is written after all
-        // object data, it means that object is in data buffer completely.
-        bufferStream.mark()
-    }
-
-    // Cut string to half a buffer size
-    override fun writeString(value: String?): Int {
-        val trimmedValue = if ((value?.length ?: 0) > MAX_STRING_SIZE) value?.substring(0, MAX_STRING_SIZE) else value
-        return super.writeString(trimmedValue)
-    }
-
-    fun mark() {
-        bufferStream.mark()
-    }
-
-    fun rollback() {
-        resetTracepointState()
-        bufferStream.rollback()
-    }
-
-    fun flush() {
-        val logicalStart = currentStartDataPosition
-        currentStartDataPosition += bufferStream.position()
-        // Don't allocate new index if it is empty
-        val indexToSave = if (index.isEmpty()) {
-            emptyList() // Singleton
-        } else {
-            val oldIndex = index
-            index = mutableListOf()
-            oldIndex
-        }
-        storage.saveDataAndIndexBlock(writerId, logicalStart, bufferStream.detachBuffer(), indexToSave)
-    }
-
-    private fun maybeFlushData() {
-        if (bufferStream.available() < DATA_FLUSH_THRESHOLD) {
-            flush()
-        }
-    }
-}
-
-private class SimpleContextSavingState: ContextSavingState {
-    private var seenClassDescriptors = BooleanArray(1024)
-    private var seenMethodDescriptors = BooleanArray(1024)
-    private var seenFieldDescriptors = BooleanArray(1024)
-    private var seenVariableDescriptors = BooleanArray(1024)
-    private var seenCodeLocations = BooleanArray(65536)
-    private val stringCache = Enumerator<String>()
-    private val accessPathCache = Enumerator<AccessPath>()
-
-    private class Enumerator<T : Any> {
-        private val cache = mutableMapOf<T, Int>()
-
-        fun isSaved(value: T): Int {
-            val id = cache[value]
-            if (id != null && id > 0) {
-                return id
-            }
-            cache[value] = -(cache.size + 1)
-            return -cache.size
-        }
-
-        fun makeSaved(value: T) {
-            val id = cache[value]
-            if (id != null && id < 0) {
-                cache[value] = -id
-            }
-        }
-    }
-
-    override fun isClassDescriptorSaved(id: Int): Boolean {
-        return id < seenClassDescriptors.size && seenClassDescriptors[id]
-    }
-
-    override fun markClassDescriptorSaved(id: Int) {
-        seenClassDescriptors = ensureSize(seenClassDescriptors, id)
-        seenClassDescriptors[id] = true
-    }
-
-    override fun isMethodDescriptorSaved(id: Int): Boolean {
-        return id < seenMethodDescriptors.size && seenMethodDescriptors[id]
-    }
-
-    override fun markMethodDescriptorSaved(id: Int) {
-        seenMethodDescriptors = ensureSize(seenMethodDescriptors, id)
-        seenMethodDescriptors[id] = true
-    }
-
-    override fun isFieldDescriptorSaved(id: Int): Boolean {
-        return id < seenFieldDescriptors.size && seenFieldDescriptors[id]
-    }
-
-    override fun markFieldDescriptorSaved(id: Int) {
-        seenFieldDescriptors = ensureSize(seenFieldDescriptors, id)
-        seenFieldDescriptors[id] = true
-    }
-
-    override fun isVariableDescriptorSaved(id: Int): Boolean {
-        return id < seenVariableDescriptors.size && seenVariableDescriptors[id]
-    }
-
-    override fun markVariableDescriptorSaved(id: Int) {
-        seenVariableDescriptors = ensureSize(seenVariableDescriptors, id)
-        seenVariableDescriptors[id] = true
-    }
-
-    override fun isCodeLocationSaved(id: Int): Boolean {
-        return id < seenCodeLocations.size && seenCodeLocations[id]
-    }
-
-    override fun markCodeLocationSaved(id: Int) {
-        seenCodeLocations = ensureSize(seenCodeLocations, id)
-        seenCodeLocations[id] = true
-    }
-
-    override fun isStringSaved(value: String): Int {
-        return stringCache.isSaved(value)
-    }
-
-    override fun markStringSaved(value: String) {
-        stringCache.makeSaved(value)
-    }
-
-    override fun isAccessPathSaved(value: AccessPath): Int {
-        return accessPathCache.isSaved(value)
-    }
-
-    override fun markAccessPathSaved(value: AccessPath) {
-        accessPathCache.makeSaved(value)
-    }
-
-    private fun ensureSize(map: BooleanArray, id: Int): BooleanArray {
-        if (id < map.size) return map
-        val newlen = Integer.max(id + 16, map.size + map.size / 2)
-        return map.copyOf(newlen)
-    }
-}
-
-private class DirectTraceWriter(
-    dataStream: OutputStream,
-    indexStream: OutputStream,
-    context: TraceContext,
-    private val pos: PositionCalculatingOutputStream = PositionCalculatingOutputStream(dataStream),
-) : TraceWriterBase(
-    context = context,
-    contextState = SimpleContextSavingState(),
-    dataStream = pos,
-    dataOutput = DataOutputStream(pos)
-) {
-    private val index = DataOutputStream(indexStream)
-
-    private var currentWriterId: Int = 0
-    private var currentBlockStart: Long = 0 // with header
-    private var currentDataStart: Long = 0 // after header
-
-    override val currentDataPosition: Long get() = pos.currentPosition - currentDataStart
-    override val writerId: Int get() = currentWriterId
-
-    init {
-        dataOutput.writeLong(TRACE_MAGIC)
-        dataOutput.writeLong(TRACE_VERSION)
-
-        index.writeLong(INDEX_MAGIC)
-        index.writeLong(TRACE_VERSION)
-    }
-
-    override fun close() {
-        super.close()
-        index.writeLong(INDEX_MAGIC)
-        index.close()
-    }
-
-    override fun writeIndexCell(kind: ObjectKind, id: Int, startPos: Long, endPos: Long) {
-        if (kind == ObjectKind.TRACEPOINT) {
-            index.writeIndexCell(kind, id, startPos, endPos)
-        } else {
-            index.writeIndexCell(kind, id, startPos + currentDataStart, endPos + currentDataStart)
-        }
-    }
-
-    fun startNewRoot(id: Int) {
-        currentWriterId = id
-        currentBlockStart = pos.currentPosition
-        dataOutput.writeKind(ObjectKind.BLOCK_START)
-        dataOutput.writeInt(id)
-        currentDataStart = pos.currentPosition
-    }
-
-    fun endRoot() {
-        val endPos = pos.currentPosition
-        dataOutput.writeKind(ObjectKind.BLOCK_END)
-        index.writeIndexCell(ObjectKind.BLOCK_START, currentWriterId, currentBlockStart, endPos)
-    }
-}
-
-class MemoryTraceCollecting(private val context: TraceContext): TraceCollectingStrategy {
-    override fun registerCurrentThread(threadId: Int) {
-        context.setThreadName(threadId, Thread.currentThread().name)
-    }
-
-    override fun completeThread(thread: Thread) {}
-
-    override fun tracePointCreated(
-        parent: TRContainerTracePoint?,
-        created: TRTracePoint
-    ) {
-        parent?.addChild(created)
-    }
-
-    override fun completeContainerTracePoint(thread: Thread, container: TRContainerTracePoint) {}
-
-    /**
-     * Do nothing.
-     * Trace collected in memory can be saved by external means, if needed.
-     */
-    override fun traceEnded() {}
-}
-
-class NullTraceCollecting(private val context: TraceContext): TraceCollectingStrategy {
-    val points = AtomicLong(0)
-
-    override fun registerCurrentThread(threadId: Int) {
-        context.setThreadName(threadId, Thread.currentThread().name)
-    }
-
-    override fun completeThread(thread: Thread) {}
-
-    override fun tracePointCreated(
-        parent: TRContainerTracePoint?,
-        created: TRTracePoint
-    ) {
-        points.incrementAndGet()
-    }
-
-    override fun completeContainerTracePoint(thread: Thread, container: TRContainerTracePoint) {}
-
-    /**
-     * Do nothing.
-     * Trace collected in memory can be saved by external means, if needed.
-     */
-    override fun traceEnded() {
-        Logger.info { "Collected ${points.get()} points" }
-    }
-}
-
-private class FileStreamingThread(
-    dataStream: OutputStream,
-    indexStream: OutputStream,
-): Thread() {
-    private sealed class Job()
-    private data class SaveBlockJob(val writerId: Int, val logicalBlockStart: Long, val dataBlock: ByteBuffer, val indexList: List<IndexCell>): Job()
-    private class ExitJob(): Job()
-
-    private val pos: PositionCalculatingOutputStream = PositionCalculatingOutputStream(dataStream)
-    private val data: OutputStream = pos
-    private val index: OutputStream = indexStream.buffered(OUTPUT_BUFFER_SIZE)
-
-    private var bufferSize = 1024
-    private var buffer: ByteBuffer = ByteBuffer.allocate(INDEX_CELL_SIZE * bufferSize)
-
-    private val queue = ArrayBlockingQueue<Job>(1024)
-
-    init {
-        name = "TR-Block-Writer"
-
-        data.writeLong(TRACE_MAGIC)
-        data.writeLong(TRACE_VERSION)
-
-        index.writeLong(INDEX_MAGIC)
-        index.writeLong(TRACE_VERSION)
-    }
-
-    fun addBlock(writerId: Int, logicalBlockStart: Long, dataBlock: ByteBuffer, indexList: List<IndexCell>) {
-        queue.put(SaveBlockJob(writerId, logicalBlockStart, dataBlock, indexList))
-    }
-
-    fun exit() {
-        queue.put(ExitJob())
-        join()
-    }
-
-    override fun run() {
-        while (true) {
-            val j = queue.take()
-            when (j) {
-                is SaveBlockJob -> saveBlock(j)
-                is ExitJob -> {
-                    closeStreams()
-                    break
-                }
-            }
-        }
-    }
-
-    private fun saveBlock(job: SaveBlockJob) {
-        val startPosition = pos.currentPosition
-        data.writeKind(ObjectKind.BLOCK_START)
-        data.writeInt(job.writerId)
-        val indexShift = pos.currentPosition - job.logicalBlockStart
-        data.write(job.dataBlock.array(), 0, job.dataBlock.limit())
-        val endPosition = pos.currentPosition
-        data.writeKind(ObjectKind.BLOCK_END)
-        data.flush()
-
-        if (bufferSize < job.indexList.size + 1) {
-            bufferSize = max(job.indexList.size + 1, (bufferSize * 1.5).toInt())
-            buffer = ByteBuffer.allocate(INDEX_CELL_SIZE * bufferSize)
-        } else {
-            buffer.clear()
-        }
-
-        buffer.putIndexCell(ObjectKind.BLOCK_START, job.writerId, startPosition, endPosition)
-        job.indexList.forEach {
-            if (it.kind == ObjectKind.TRACEPOINT) {
-                // Trace points indices are in "local" offsets, as they should be loaded
-                // from interleaving per-thread blocks
-                buffer.putIndexCell(it.kind, it.id, it.startPos, it.endPos)
-            } else {
-                // All other objects are in "global" offsets as they cannot be split between blocks
-                // and can be loaded without taking blocks in consideration
-                buffer.putIndexCell(it.kind, it.id, it.startPos + indexShift, it.endPos + indexShift)
-            }
-        }
-        buffer.flip()
-        index.write(buffer.array(), 0, buffer.limit())
-        index.flush()
-    }
-
-    private fun closeStreams() {
-        data.writeKind(ObjectKind.EOF)
-        data.close()
-
-        buffer.clear()
-        buffer.putIndexCell(ObjectKind.EOF, -1, -1, -1)
-        buffer.flip()
-        index.write(buffer.array(), 0, buffer.limit())
-        index.writeLong(INDEX_MAGIC)
-        index.close()
-    }
-
-    private companion object {
-        private val staticBuffer = ByteBuffer.allocate(Long.SIZE_BYTES)
-
-        private fun OutputStream.writeKind(v: ObjectKind) {
-            write(v.ordinal)
-        }
-
-        private fun OutputStream.writeInt(v: Int) {
-            staticBuffer.clear()
-            staticBuffer.putInt(v)
-            staticBuffer.flip()
-            write(staticBuffer.array(), 0, Int.SIZE_BYTES)
-        }
-
-        private fun OutputStream.writeLong(v: Long) {
-            staticBuffer.clear()
-            staticBuffer.putLong(v)
-            staticBuffer.flip()
-            write(staticBuffer.array(), 0, Long.SIZE_BYTES)
-        }
-
-        private fun ByteBuffer.putKind(v: ObjectKind) {
-            put(v.ordinal.toByte())
-        }
-    }
-}
-
-class FileStreamingTraceCollecting(
-    dataStream: OutputStream,
-    indexStream: OutputStream,
-    val context: TraceContext
-): TraceCollectingStrategy, ContextSavingState {
-    constructor(baseFileName: String, context: TraceContext) :
-            this(
-                dataStream = openNewFile(baseFileName),
-                indexStream = openNewFile("$baseFileName.$INDEX_FILENAME_EXT"),
-                context = context
-            )
-
-    private val ioThread = FileStreamingThread(dataStream, indexStream)
-    init {
-        ioThread.start()
-    }
-
-    private var seenClassDescriptors = AtomicBitmap()
-    private var seenMethodDescriptors = AtomicBitmap()
-    private var seenFieldDescriptors = AtomicBitmap()
-    private var seenVariableDescriptors = AtomicBitmap()
-    private var seenCodeLocations = AtomicBitmap()
-    private val stringEnumerator = Enumerator<String>()
-    private val accessPathEnumerator = Enumerator<AccessPath>()
-
-    private val writers = ConcurrentHashMap<Thread, BufferedTraceWriter>()
-
-    private class Enumerator<T : Any> {
-        private val idGenerator = AtomicInteger(1)
-        private val cache = ConcurrentHashMap<T, Int>()
-
-        fun isSaved(value: T): Int {
-            val id = cache.computeIfAbsent(value) { _ -> -idGenerator.getAndIncrement() }
-            return id
-        }
-
-        fun makeSaved(value: T) {
-            // Make positive!
-            cache.compute(value) { _, v -> v?.absoluteValue }
-        }
-    }
-
-    override fun registerCurrentThread(threadId: Int) {
-        val thread = Thread.currentThread()
-        if (writers[thread] != null) return
-        writers[thread] = BufferedTraceWriter(
-            writerId = threadId,
-            context = context,
-            contextState = this,
-            // This is needed to work around visibility problems
-            storage = object : BlockSaver {
-                override fun saveDataAndIndexBlock(writerId: Int, logicalBlockStart: Long, dataBlock: ByteBuffer, indexList: List<IndexCell>) =
-                    ioThread.addBlock(writerId, logicalBlockStart, dataBlock, indexList)
-            }
-        )
-        context.setThreadName(threadId, thread.name)
-    }
-
-    override fun completeThread(thread: Thread) {
-        val writer = writers[thread] ?: return
-        writer.flush()
-    }
-
-    override fun tracePointCreated(
-        parent: TRContainerTracePoint?,
-        created: TRTracePoint
-    ) {
-        val writer = writers[Thread.currentThread()] ?: return
-        try {
-            writer.mark()
-            created.save(writer)
-        } catch (_: BufferOverflowException) {
-            // Flush current buffers, start over
-            writer.rollback()
-            writer.flush()
-            created.save(writer)
-        }
-    }
-
-    override fun completeContainerTracePoint(thread: Thread, container: TRContainerTracePoint) {
-        val writer = writers[thread] ?: return
-        try {
-            writer.mark()
-            container.saveFooter(writer)
-        } catch (_: BufferOverflowException) {
-            // Flush current buffers, start over
-            writer.rollback()
-            writer.flush()
-            container.saveFooter(writer)
-        }
-    }
-
-    override fun traceEnded() {
-        writers.entries.forEach { (t, w) ->
-            // writerId matches the assigned id of the thread
-            w.writeThreadName(w.writerId, t.name)
-        }
-        writers.values.forEach { it.close() }
-
-        // Flush all output & exit
-        ioThread.exit()
-    }
-
-    override fun isClassDescriptorSaved(id: Int): Boolean = seenClassDescriptors.isSet(id)
-
-    override fun markClassDescriptorSaved(id: Int): Unit = seenClassDescriptors.set(id)
-
-    override fun isMethodDescriptorSaved(id: Int): Boolean = seenMethodDescriptors.isSet(id)
-
-    override fun markMethodDescriptorSaved(id: Int): Unit = seenMethodDescriptors.set(id)
-
-    override fun isFieldDescriptorSaved(id: Int): Boolean = seenFieldDescriptors.isSet(id)
-
-    override fun markFieldDescriptorSaved(id: Int): Unit = seenFieldDescriptors.set(id)
-
-    override fun isVariableDescriptorSaved(id: Int): Boolean = seenVariableDescriptors.isSet(id)
-
-    override fun markVariableDescriptorSaved(id: Int): Unit = seenVariableDescriptors.set(id)
-
-    override fun isCodeLocationSaved(id: Int): Boolean = seenCodeLocations.isSet(id)
-
-    override fun markCodeLocationSaved(id: Int): Unit = seenCodeLocations.set(id)
-
-    override fun isStringSaved(value: String): Int = stringEnumerator.isSaved(value)
-
-    override fun markStringSaved(value: String): Unit = stringEnumerator.makeSaved(value)
-
-    override fun isAccessPathSaved(value: AccessPath): Int = accessPathEnumerator.isSaved(value)
-
-    override fun markAccessPathSaved(value: AccessPath): Unit = accessPathEnumerator.makeSaved(value)
-}
-
-/**
- * Top-level function to save full-depth recorded trace old-style (all in once)
- */
-fun saveRecorderTrace(baseFileName: String, context: TraceContext, rootCallsPerThread: List<TRTracePoint>) =
-    saveRecorderTrace(
-        data = openNewFile(baseFileName).buffered(OUTPUT_BUFFER_SIZE),
-        index = openNewFile("$baseFileName.$INDEX_FILENAME_EXT").buffered(OUTPUT_BUFFER_SIZE),
-        context = context,
-        rootCallsPerThread = rootCallsPerThread
-    )
-
-fun saveRecorderTrace(data: OutputStream, index: OutputStream, context: TraceContext, rootCallsPerThread: List<TRTracePoint>) {
-    check(context == TRACE_CONTEXT) { "Now only global TRACE_CONTEXT is supported" }
-
-    DirectTraceWriter(data, index, context).use { tw ->
-        rootCallsPerThread.forEachIndexed { id, root ->
-            tw.startNewRoot(id)
-            tw.writeThreadName(id, context.getThreadName(id))
-            saveTRTracepoint(tw, root)
-            tw.endRoot()
-        }
-    }
-}
-
-fun packRecordedTrace(baseFileName: String, metaInfo: TraceMetaInfo, deleteSources: Boolean = true) {
-    val dataName = baseFileName
-    val indexName = "$baseFileName.$INDEX_FILENAME_EXT"
-    val outputName = "$baseFileName.$PACK_FILENAME_EXT"
-
-    try {
-        ZipOutputStream(openNewFile(outputName).buffered(OUTPUT_BUFFER_SIZE)).use { zip ->
-            var input: InputStream
-
-            // Don't compress for now, but STORED requires pre-compute CRC32, and we don't want to do it.
-            zip.setMethod(ZipOutputStream.DEFLATED)
-            zip.setLevel(0)
-
-            // Store metadata first
-            zip.putNextEntry(ZipEntry(PACKED_META_ITEM_NAME))
-            val printer = PrintWriter(zip)
-            metaInfo.print(printer)
-            printer.flush() // Don't close, it will close ZIP stream too!
-            zip.closeEntry()
-
-            // Store data
-            zip.putNextEntry(ZipEntry(PACKED_DATA_ITEM_NAME))
-            input = openExistingFile(dataName) ?: throw IllegalArgumentException("Cannot open trace file \"$dataName\"")
-            input.use { data ->
-                data.copyTo(zip)
-            }
-            zip.closeEntry()
-
-            // Store Index
-            zip.putNextEntry(ZipEntry(PACKED_INDEX_ITEM_NAME))
-            input = openExistingFile(indexName) ?: throw IllegalArgumentException("Cannot open trace index file \"$indexName\"")
-            input.use { data ->
-                data.copyTo(zip)
-            }
-            zip.closeEntry()
-        }
-        if (deleteSources) {
-            Files.deleteIfExists(Path(dataName))
-            Files.deleteIfExists(Path(indexName))
-        }
-    } catch (e: Throwable) {
-        Files.deleteIfExists(Path(outputName))
-        throw e
-    }
-}
-
-private fun saveTRTracepoint(writer: TraceWriter, tracepoint: TRTracePoint) {
-    tracepoint.save(writer)
-    if (tracepoint is TRContainerTracePoint) {
-        tracepoint.events.forEach {
-            if (it != null) {
-                saveTRTracepoint(writer, it)
-            }
-        }
-        tracepoint.saveFooter(writer)
-    }
-}
-
-private fun DataOutput.writeIndexCell(kind: ObjectKind, id: Int, startPos: Long, endPos: Long) {
-    writeByte(kind.ordinal)
-    writeInt(id)
-    writeLong(startPos)
-    writeLong(endPos)
-}
-
-private fun ByteBuffer.putIndexCell(kind: ObjectKind, id: Int, startPos: Long, endPos: Long) {
-    put(kind.ordinal.toByte())
-    putInt(id)
-    putLong(startPos)
-    putLong(endPos)
-}
